+++ src/java/quickfix_ScreenLogFactory.cpp
@@ -78,7 +78,7 @@ JNIEXPORT jobject JNICALL Java_quickfix_ScreenLogFactory_create__Lquickfix_Sessi
     FIX::Log* pLog = pFactory->create( *pSessionID );
     JVMClass type( "Lquickfix/ScreenLog;" );
     jmethodID method = pEnv->GetMethodID( type, "<init>", "(J)V" );
-    JVMObject result = pEnv->NewObject( type, method, ( jint ) pLog );
+    JVMObject result = pEnv->NewObject( type, method, ( jlong ) pLog );
     return result;
   }
   catch ( FIX::ConfigError & e )
